+++ tests/test_power_expr_simple.m
@@ -23,6 +23,6 @@ function test_power_expr_simple
     B = A.power_expr(2);
     fprintf('B  = %s\n', B.toString());
     fprintf('B_ = %s\n', B_.toString());
-    assert(norm(B.expr - B_.expr) == 0);    
+    assert(norm(double(B.expr - B_.expr)) == 0);    
 end
